changing to not use deep learning model and rewrite the weight.py
Undo to not train the DL because the sequence could not implemented by the condition of the training. The LSTM must be sequential but some of the label appear only in the bottom of the dataset. Thus the split 50:50 cannot be done. Also, the random are violate the LSTM that must input in sequence. Therefore, the sequential cannot be trained.

The weight.py fixing into server backend and frontend.